requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt --resolver=backtracking requirements-dev.in setup.py
#
astroid==2.15.4
    # via pylint
black==23.3.0
    # via -r requirements-dev.in
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   typer
coverage==7.2.5
    # via -r requirements-dev.in
dill==0.3.6
    # via pylint
exceptiongroup==1.1.1
    # via pytest
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
fsspec==2023.4.0
    # via huggingface-hub
huggingface-hub==0.14.1
    # via
    #   imaginairy-normal-map (setup.py)
    #   timm
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via
    #   -r requirements-dev.in
    #   pylint
jinja2==3.1.2
    # via torch
lazy-object-proxy==1.9.0
    # via astroid
libcst==0.4.9
    # via pycln
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via
    #   pylama
    #   pylint
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
networkx==3.1
    # via torch
numpy==1.24.3
    # via torchvision
packaging==23.1
    # via
    #   black
    #   huggingface-hub
    #   pytest
    #   pytest-sugar
pathspec==0.10.3
    # via
    #   black
    #   pycln
pillow==9.5.0
    # via torchvision
platformdirs==3.5.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
pycln==2.1.3
    # via -r requirements-dev.in
pycodestyle==2.10.0
    # via pylama
pydocstyle==6.3.0
    # via pylama
pyflakes==3.0.1
    # via pylama
pylama==8.4.1
    # via -r requirements-dev.in
pylint==2.17.3
    # via -r requirements-dev.in
pytest==7.3.1
    # via
    #   -r requirements-dev.in
    #   pytest-randomly
    #   pytest-sugar
pytest-randomly==3.12.0
    # via -r requirements-dev.in
pytest-sugar==0.9.7
    # via -r requirements-dev.in
pyyaml==6.0
    # via
    #   huggingface-hub
    #   libcst
    #   pycln
    #   responses
    #   timm
requests==2.29.0
    # via
    #   huggingface-hub
    #   responses
    #   torchvision
responses==0.23.1
    # via -r requirements-dev.in
ruff==0.0.264
    # via -r requirements-dev.in
safetensors==0.3.1
    # via imaginairy-normal-map (setup.py)
snowballstemmer==2.2.0
    # via pydocstyle
sympy==1.11.1
    # via torch
termcolor==2.3.0
    # via pytest-sugar
timm==0.6.13
    # via imaginairy-normal-map (setup.py)
tomli==2.0.1
    # via
    #   black
    #   pylint
    #   pytest
tomlkit==0.11.8
    # via
    #   pycln
    #   pylint
torch==2.0.0
    # via
    #   imaginairy-normal-map (setup.py)
    #   timm
    #   torchvision
torchvision==0.15.1
    # via
    #   imaginairy-normal-map (setup.py)
    #   timm
tqdm==4.65.0
    # via huggingface-hub
typer==0.7.0
    # via pycln
types-pyyaml==6.0.12.9
    # via responses
typing-extensions==4.5.0
    # via
    #   astroid
    #   huggingface-hub
    #   libcst
    #   torch
    #   typing-inspect
typing-inspect==0.8.0
    # via libcst
urllib3==1.26.15
    # via
    #   requests
    #   responses
wheel==0.40.0
    # via -r requirements-dev.in
wrapt==1.15.0
    # via astroid
